Improved Education and fixed footer
Improved the Education.html by creating a div which the classes were put into. Fixed the footer to make it work better. Used some javascript to make the footer work better. Duplicated the education page to education.1.html to allow content to be moved over to the education.html page. education.1.html is temporary and will be deleted in the next push.